Restructing, to avoid app.js becoming monolithic
