+++ app.js
@@ -1,17 +1,20 @@
-var express = require('express');
-var moment = require('moment');
-var memoryCache = require('memory-cache');
-var request = require('request');
-var http = require("http");
+const express = require('express');
+const moment = require('moment');
+const memoryCache = require('memory-cache');
+const request = require('request');
+const http = require("http");
+
+const events = require('./lib/services/events');
+const occupantsDirectory = require('./lib/services/occupants-directory');
+const netatmo = require('./lib/services/netatmo');
+const mySeat = require('./lib/services/myseat');
+const vendor = require('./lib/services/vendor');
+
+const time = require('./lib/services/time');
 
 const MINUTES = 60000;
 const CACHE_TIMEOUT = 15 * MINUTES;
 const DEFAULT_TIMEZONE = 'America/Toronto';
-
-var getCalendarEvents = require('./lib/get-calendar-events');
-var adaptCalendarEvent = require('./lib/adapt-calendar-event');
-var groupEventsByDate = require('./lib/group-events-by-date');
-var getOccupants = require('./lib/occupants-directory');
 
 var app = express();
 
@@ -33,31 +36,8 @@ app.use(logger('dev'));
 
 app.use(express.static('public'));
 
-app.get('/api/keys', function(req, res, next) {
-    // TODO return all the vendor keys. limit by requesting domain?
-    res.json({
-        myseat: process.env.MYSEAT_API_KEY
-    });
-});
-
-app.get(/^\/vendor\/(.+)\/(.*)/, function(req, res, next) {
-    // TODO return all the vendor keys. limit by requesting domain?
-    console.log('param 0');
-    console.log('param 1');
-
-    var baseUrl;
-    if (req.params[0] === 'myseat') {
-        console.log('vendor myseat');
-        baseUrl = 'https://apiv3.myseat.fr/Request/'
-    }
-
-    res.json({
-        myseat: process.env.MYSEAT_API_KEY
-    });
-});
-
-
 // Route handlers
+
 
 // TODO have the version actually be the GIT hash
 app.get('/api/', function(req, res, next) {
@@ -75,130 +55,14 @@ app.get('/api/', function(req, res, next) {
     })
 });
 
-app.get('/api/events', function(req, res, next) {
 
-    // flush the cache if we get flushcache=1, added to allow for cases
-    // where the calendar was updated and we want to see the results immediately
+app.get(/^\/api\/vendor\/([^/]+)\/(.*)/, vendor.handleGets);
+app.get('/api/events', events.handleGetEvents);
+app.get('/api/netatmo/environment', netatmo.handleGetStationData);
 
-    if ('1' === req.query['flushcache']) {
-        memoryCache.clear();
-    }
-
-    // we use url as key, in order to support request with varying parameters
-    var cachedEvents = memoryCache.get(req.url);
-    if (cachedEvents !== null) {
-        res.json(cachedEvents);
-        return;
-    }
-
-    var options = {};
-
-    var twentyfourHour = req.query['24hour'];
-    var showDesc = req.query['desc'];
-    options.showDesc = (showDesc && showDesc === '1');
-    var format = (twentyfourHour && twentyfourHour === '1' ? 'HH:mm' : 'h:mm a');
-    options.timeFormat = format;
-
-    // handle option of returning time in UTC
-    var utc = req.query['utc'];
-    var tz = (utc && utc === '1' ? 'UTC' : DEFAULT_TIMEZONE);
-    options.timezone = tz;
-
-    var calendarId = process.env.GOOGLE_CALENDAR_ID;
-    var apiKey = process.env.GOOGLE_API_KEY;
-
-    if (!calendarId || !apiKey) {
-        throw new Error('Events API requires calendar config.');
-    }
-
-    getCalendarEvents(calendarId, apiKey).then(function(events) {
-        var i;
-        if ('1' !== req.query['private']) {
-            var selectedEvents = [];
-
-            for (i = 0; i < events.length; i++) {
-                if (events[i].summary.endsWith('*')) {
-                    selectedEvents.push(events[i]);
-                }
-            }
-            events = selectedEvents;
-        }
-
-        var adaptedEvents = events.map(
-            event => adaptCalendarEvent(event, options)
-        ).slice(0, 5);
-
-        var groupedEvents = groupEventsByDate(adaptedEvents);
-        res.json(groupedEvents);
-        memoryCache.put(req.url, groupedEvents, CACHE_TIMEOUT);
-    }).catch(next);
-});
-
-const getNetatmoEnvironment = require('./lib/get-netatmo-environment');
-app.get('/api/netatmo/environment', function(req, res, next) {
-    getNetatmoEnvironment(
-        process.env.NETATMO_CLIENT_ID,
-        process.env.NETATMO_CLIENT_SECRET,
-        process.env.NETATMO_USERNAME,
-        process.env.NETATMO_PASSWORD,
-        function(netatmoEnvironment) {
-            res.send(netatmoEnvironment);
-        });
-});
-
-/**
- * Returns the time on the server. Provided, since the time
- * on the Raspberry Pi was proving unreliable. 
- * 
- */
-app.get('/api/time', function(req, res, next) {
-    // handle option of returning time in 24 hours
-    var twentyfourHour = req.query['24hour'];
-    var format = (twentyfourHour && twentyfourHour === '1' ? 'HH:mm' : 'h:mm a');
-    // handle option of returning time in UTC
-    var utc = req.query['utc'];
-    var tz = (utc && utc === '1' ? 'UTC' : DEFAULT_TIMEZONE);
-    var timeString = moment().tz(tz).format(format);
-    res.json({ time: timeString });
-});
-
-/**
- * Returns the occupant directory of Notman House. Currently
- * only deals with businesses names and Notman management staff.
- * 
- * Currently accepted query parameters:
- *   - floor
- *   - building
- */
-app.get('/api/directory', function(req, res, next) {
-    var floor = req.query.floor;
-    var building = req.query.building;
-    var format = req.query.format;
-    if (!format || format === 'json') {
-        res.json(getOccupants(floor, building, format));
-    } else {
-        res.type('text/plain');
-        res.send(getOccupants(floor, building, format));
-    }
-})
-
-/**
- * Returns the mySeat data. Should this be removed due to mySeat
- * no longer being part of the 'Hack the House' initiative?
- */
-app.get('/api/myseat/chairs', function(req, res, next) {
-    var url = `https://apiv3.myseat.fr/Request/GetChairs/key/${process.env.MYSEAT_API_KEY}`;
-
-    request(url, function(err, result) {
-        if (err) {
-            res.status(500).send();
-            console.error(err);
-            return;
-        }
-
-        res.status(result.statusCode).type(result.headers['content-type']).send(result.body);
-    });
-});
+app.get('/api/time', time.handleGetTime);
+app.get('/api/directory', occupantsDirectory.handleGetOccupants);
+app.get('/api/myseat/chairs', mySeat.handleGetChairs);
 
 app.post('/refresh', function(req, res, next) {
     app.wss.clients.forEach(ws => {
