+++ lib/config/occupants.js
@@ -1,35 +1,35 @@
 module.exports = [
-       { room: 101, floor: 1, building: 'campus', occupant: 'With Intent Marketing Services' },
-       { room: 102, floor: 1, building: 'campus', occupant: 'PayHero' },
-       { room: 103, floor: 1, building: 'campus', occupant: 'FundMetric (FF)' },
-       { room: 104, floor: 1, building: 'campus', occupant: 'Kids Code Jeunesse' },
-       { room: 105, floor: 1, building: 'campus', occupant: 'The Bubbles Company' },
-       { room: 106, floor: 1, building: 'campus', occupant: 'Aspertise Inc.' },
-       { room: 107, floor: 1, building: 'campus', occupant: 'Corstem Inc.' },
-       { room: 108, floor: 1, building: 'campus', occupant: '-' },
-       { room: 109, floor: 1, building: 'campus', occupant: 'DogSync' },
-       { room: 110, floor: 1, building: 'campus', occupant: '-' },
-       { room: 111, floor: 1, building: 'campus', occupant: 'Meeting Room' },
-       { room: 112, floor: 1, building: 'campus', occupant: 'Octoraise' },
+    { room: 101, floor: 1, building: 'campus', occupant: 'With Intent Marketing Services' },
+    { room: 102, floor: 1, building: 'campus', occupant: 'PayHero' },
+    { room: 103, floor: 1, building: 'campus', occupant: 'FundMetric (FF)' },
+    { room: 104, floor: 1, building: 'campus', occupant: 'Kids Code Jeunesse' },
+    { room: 105, floor: 1, building: 'campus', occupant: 'The Bubbles Company' },
+    { room: 106, floor: 1, building: 'campus', occupant: 'Aspertise Inc.' },
+    { room: 107, floor: 1, building: 'campus', occupant: 'Corstem Inc.' },
+    { room: 108, floor: 1, building: 'campus', occupant: '-' },
+    { room: 109, floor: 1, building: 'campus', occupant: 'DogSync' },
+    { room: 110, floor: 1, building: 'campus', occupant: '-' },
+    { room: 111, floor: 1, building: 'campus', occupant: 'Meeting Room' },
+    { room: 112, floor: 1, building: 'campus', occupant: 'Octoraise' },
 
-       { room: 201, floor: 2, building: 'campus', occupant: 'Les Technologies Discoshare Inc.' },
-       { room: 202, floor: 2, building: 'campus', occupant: '[Meeting Room]' },
-       { room: 203, floor: 2, building: 'campus', occupant: 'Groupe Autour du Monde' },
-       { room: 204, floor: 2, building: 'campus', occupant: 'Macromeasures' },
-       { room: 205, floor: 2, building: 'campus', occupant: 'Taste Bay' },
-       { room: 206, floor: 2, building: 'campus', occupant: 'Prollster (FF)' },
-       { room: 207, floor: 2, building: 'campus', occupant: 'Men at Heights' },
-       { room: 208, floor: 2, building: 'campus', occupant: 'Guides (FF)' },
-       { room: 209, floor: 2, building: 'campus', occupant: 'MotorLeaf (FF)' },
-       { room: 210, floor: 2, building: 'campus', occupant: 'Nabes' },
-       { room: 211, floor: 2, building: 'campus', occupant: 'Notman: Ritika Dutt' },
-       { room: 212, floor: 2, building: 'campus', occupant: 'Keatext' },
-       { room: 213, floor: 2, building: 'campus', occupant: 'Notman: Emma Williams' },
+    { room: 201, floor: 2, building: 'campus', occupant: 'Les Technologies Discoshare Inc.' },
+    { room: 202, floor: 2, building: 'campus', occupant: '[Meeting Room]' },
+    { room: 203, floor: 2, building: 'campus', occupant: 'Groupe Autour du Monde' },
+    { room: 204, floor: 2, building: 'campus', occupant: 'Macromeasures' },
+    { room: 205, floor: 2, building: 'campus', occupant: 'Taste Bay' },
+    { room: 206, floor: 2, building: 'campus', occupant: 'Prollster (FF)' },
+    { room: 207, floor: 2, building: 'campus', occupant: 'Men at Heights' },
+    { room: 208, floor: 2, building: 'campus', occupant: 'Guides (FF)' },
+    { room: 209, floor: 2, building: 'campus', occupant: 'MotorLeaf (FF)' },
+    { room: 210, floor: 2, building: 'campus', occupant: 'Nabes' },
+    { room: 211, floor: 2, building: 'campus', occupant: 'Notman: Ritika Dutt' },
+    { room: 212, floor: 2, building: 'campus', occupant: 'Keatext' },
+    { room: 213, floor: 2, building: 'campus', occupant: 'Notman: Emma Williams' },
 
-       { room: 1, floor: 1, building: 'front', occupant: 'Real Ventures' },
-       { room: 2, floor: 1, building: 'front', occupant: 'Flow Ventures' },
-       { room: 3, floor: 2, building: 'front', occupant: 'Innocité' },
-       { room: 4, floor: 2, building: 'front', occupant: 'Desjardins' },
-       { room: 5, floor: 2, building: 'front', occupant: 'Montreal Newtech' },
-       { room: 6, floor: 2, building: 'front', occupant: 'Pyxis' }
-   ];
+    { room: 1, floor: 1, building: 'front', occupant: 'Real Ventures' },
+    { room: 2, floor: 1, building: 'front', occupant: 'Flow Ventures' },
+    { room: 3, floor: 2, building: 'front', occupant: 'Innocité' },
+    { room: 4, floor: 2, building: 'front', occupant: 'Desjardins' },
+    { room: 5, floor: 2, building: 'front', occupant: 'Montreal Newtech' },
+    { room: 6, floor: 2, building: 'front', occupant: 'Pyxis' }
+];
