+++ lib/services/events/adapt-calendar-event.js
@@ -1,6 +1,6 @@
-var moment = require('moment-timezone');
-var parseStringTags = require('./parse-string-tags');
-var clockDurationBetween = require('./clock-duration-between');
+const moment = require('moment-timezone');
+const parseStringTags = require('./parse-string-tags');
+const clockDurationBetween = require('./clock-duration-between');
 
 module.exports = function adaptCalendarEvent(event, options) {
