+++ lib/services/events/group-events-by-date.js
