+++ lib/services/netatmo/get-netatmo-environment.js
@@ -11,22 +11,22 @@ const netatmo = require('netatmo');
  * @param {function} callback The function to call on completion.
  * @returns {Array}
  */
-module.exports = function getNetatmoEnvironment(clientId, clientSecret,
-                                                username, password, callback) {
-  var auth = {
-    "client_id": clientId,
-    "client_secret": clientSecret,
-    "username": username,
-    "password": password
-  };
+module.exports.getStationsData = function getNetatmoEnvironment(clientId, clientSecret,
+    username, password, callback) {
+    var auth = {
+        "client_id": clientId,
+        "client_secret": clientSecret,
+        "username": username,
+        "password": password
+    };
 
-  var api = new netatmo(auth);
+    var api = new netatmo(auth);
 
-  api.getDevicelist(function(err, devices, modules) {
-    if(err) {
-      console.error(err);
-      return callback([]);
-    }
-    return callback(devices.concat(modules));
-  });
+    api.getStationsData(function(err, devices, modules) {
+        if (err) {
+            console.error(err);
+            return callback([]);
+        }
+        return callback(devices.concat(modules));
+    });
 };
